requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dash==2.9.1
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.2
fonttools==4.38.0
gevent==22.10.2
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
itsdangerous==2.1.2
Jinja2==3.1.2
kaleido==0.2.1
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pkg_resources==0.0.0
plotly==5.13.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
tenacity==8.2.1
urllib3==2.0.2
uWSGI==2.0.21
Werkzeug==2.2.3
zipp==3.13.0
zope.event==4.6
zope.interface==5.5.2
